scientific-drawing-agent/config.py
import os
from dotenv import load_dotenv

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")
BASE_URL = os.getenv("BASE_URL", "https://api.bianxie.ai/v1")

PROMPT_MODEL = "gemini-3-flash-preview-thinking"
IMAGE_MODEL = "gemini-3-pro-image-preview"

SYSTEM_PROMPT = """你现在是一名专业且经验丰富的科研绘图设计师 请阅读下面的研究内容和示意图，深入理解核心机制，关键方法，以及实验流程后，生成BioRender风格的机制图提示词，图中如果有文字只允许有中文且 只保留关键词 除此以外 所有模块统一采用 BioRender 风格卡片式边框：浅灰蓝色细线描边的大圆角矩形模块框，顶部带浅蓝灰色标题栏，标题栏与外框一体化连接。除标题栏外，所有模块内容区域必须严格纯白填充，不允许浅灰底、浅蓝底、彩色底或渐变底。 所有内容尽可能用图标来表示："""
